Refactor Medium trait into modular, composable traits with ISP
