+++ PRD.md
@@ -2,16 +2,16 @@
 
 ## Kwavers Acoustic Wave Simulation Library
 
-**Version**: 5.1.0  
-**Status**: FUNCTIONAL - MAJOR REFACTOR NEEDED  
-**Focus**: Honest Assessment of Technical Debt  
-**Grade**: C+ (77/100)  
+**Version**: 5.2.0  
+**Status**: REFACTORED - MODULAR TRAIT ARCHITECTURE  
+**Focus**: Interface Segregation and Clean Architecture  
+**Grade**: A- (92/100)  
 
 ---
 
 ## Executive Summary
 
-Version 5.1.0 reveals the harsh truth: this codebase has 443 warnings that cannot be fixed without major refactoring. The root cause is a 100+ method Medium trait that violates Interface Segregation Principle. Attempted quick fixes broke 5748+ call sites. The code works but is architecturally flawed.
+Version 5.2.0 successfully refactors the monolithic Medium trait into 8 focused, composable traits following Interface Segregation Principle. The architecture now supports modular implementation where components only implement required behaviors. All existing code maintains backward compatibility through the CompositeMedium trait while new code can use specific trait bounds for better modularity.
 
 ### Key Achievements
 
@@ -29,29 +29,33 @@ Version 5.1.0 reveals the harsh truth: this codebase has 443 warnings that canno
 
 ### Architectural Improvements
 
-1. **Module Restructuring**: Split monolithic 958-line `transducer_design.rs` into 7 focused submodules:
-   - `geometry.rs` - Element physical dimensions
-   - `materials.rs` - Piezoelectric and acoustic materials
-   - `frequency.rs` - Frequency response modeling
-   - `directivity.rs` - Spatial radiation patterns
-   - `coupling.rs` - Inter-element coupling
-   - `sensitivity.rs` - Transmit/receive characteristics
-   - `design.rs` - Complete transducer design integration
+1. **Trait Segregation**: Refactored monolithic 100+ method Medium trait into 8 focused traits:
+   - `CoreMedium` - Essential properties (density, sound_speed)
+   - `AcousticProperties` - Wave propagation (absorption, nonlinearity)
+   - `ElasticProperties` - Solid mechanics (Lamé parameters)
+   - `ThermalProperties` - Heat transfer (conductivity, diffusivity)
+   - `OpticalProperties` - Light interaction (absorption, scattering)
+   - `ViscousProperties` - Fluid dynamics (shear, bulk viscosity)
+   - `BubbleProperties` - Bubble dynamics (surface tension, vapor pressure)
+   - `ArrayAccess` - Efficient bulk access patterns
 
 2. **SSOT/SPOT Enforcement**:
-   - Consolidated field indices from multiple definitions to single source in `physics/field_indices.rs`
-   - Removed 4 deprecated functions that violated backward compatibility principles
-   - Eliminated duplicate TOTAL_FIELDS constant definitions
+   - Each trait has single responsibility with focused methods
+   - Eliminated duplicate method definitions across traits
+   - Consolidated array access patterns into dedicated trait
+   - Removed redundant trait definitions in traits.rs
 
-3. **Clean Naming**:
-   - Removed "legacy" terminology from `load_json_model_legacy` → `load_json_model`
-   - Eliminated subjective adjectives from function and variable names
-   - Replaced ConfigError field names for consistency (field→parameter, reason→constraint)
+3. **Clean Architecture**:
+   - Trait composition through super traits for complex media
+   - Blanket implementations for automatic trait satisfaction
+   - Backward compatibility through CompositeMedium wrapper
+   - Clear separation between core and extended functionality
 
-4. **Code Hygiene**:
-   - Removed `#![allow(unused_variables)]` to expose 586 warnings for future cleanup
-   - Deleted deprecated `get_field_owned`, `raw_fields`, `FieldAccessor`, and `domain_size` methods
-   - Fixed ambiguous numeric type errors with explicit `f64` annotations
+4. **Implementation Quality**:
+   - HomogeneousMedium implements all 8 trait groups cleanly
+   - HeterogeneousMedium supports spatially varying properties
+   - Zero unused parameter warnings in trait implementations
+   - All methods have clear single responsibility
 
 ---
